requirements.txt
discord.py>=2
python-dotenv
httpx
PyYAML
Pillow
